+++ utils/src/Vulkan/Utils/QueueAssignment.hs
@@ -102,7 +102,9 @@ assignQueues
   -- ^ A set of requirements for 'Queue's to be created
   -> m
        ( Maybe
-           (Vector (DeviceQueueCreateInfo '[]), Device -> n (f Queue))
+           ( Vector (DeviceQueueCreateInfo '[])
+           , Device -> n (f (QueueFamilyIndex, Queue))
+           )
        )
   -- ^
   -- - A set of 'DeviceQueueCreateInfo's to pass to 'createDevice'
@@ -177,11 +179,11 @@ assignQueues phys specs = runMaybeT $ do
         ]
 
       -- Get
-      extractQueues :: Device -> n (f Queue)
+      extractQueues :: Device -> n (f (QueueFamilyIndex, Queue))
       extractQueues dev =
         for specsWithQueueIndex
-          $ \(_, QueueFamilyIndex familyIndex, QueueIndex index) ->
-              getDeviceQueue dev familyIndex index
+          $ \(_, i@(QueueFamilyIndex familyIndex), QueueIndex index) ->
+              (i, ) <$> getDeviceQueue dev familyIndex index
 
   pure (queueCreateInfos, extractQueues)