+++ utils/changelog.md
@@ -5,6 +5,7 @@
 - Change type of pickPhysicalDevice to return Nothing instead of throwing
 - Add `checkCommandsExp` function to generate an expression checking specified
   commands for non-nullness
+- Expose Queue family index for queues assigned with `assignQueues`
 
 ## [0.2] - 2020-11-15